requirements.txt
langchain-core
langchain-openai
langchain-community
langgraph
pydantic
pypdf
python-dotenv
langsmith